Current In Progress Work
X is the temporary shortcut key to access the new routing interface.  The interface will take up to 10 paths and calculate how much time it will take to complete accounting for if it's a new game playthrough, if getting keys, if has weapons, and if cutscenes have been viewed yet.  This is the end goal, but as of this commit, only some of these are implemented.  There might be some other unrelated changes in here since I am using a more updated Unity version compared to the original project.